requirements.txt
aiogram==2.19
gino==1.0.1
aioredis==2.0.1
alembic==1.7.7
PyYAML==6.0
orjson==3.6.7
gtts==2.2.4
pytest==7.1.0
pytest-asyncio==0.18.2
freezegun==1.2.0